rename utils.py and add more util functions
